chore(runtime): use parametrized tests to test routes loaders
